Hierarchy changes
Hierarchy is completely changed.

Application is divided into sections. For example first section is
loginActivity and it't package name is  _1_loginSection
Second is changePasswordActivity,name of package is
_2_changePasswordSection  . . .  etc

All packages (which are predictable to be in project) are already
created by me. Please fill everything related to X package in the X
package. All commin files are in the common package. The are aranged as
much as possible. Please keep it like that :)

And finally , 4 methods created by Artak, are moved to
I_CommonMethodsForWorkingWithServer interface.
They are `
void getDataFromServer();
void sendDataToServer(String... strings);
void initCustomLoading();
void showError(String text);
void loading(boolean show);

These methods by this time are used in changePassowrdActivity,
vocabularyActvity and ClassroomActivity.